Introduce Lombok @Builder and @Value

+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/EthServer.java
@@ -92,7 +92,7 @@ class EthServer {
           constructGetHeadersResponse(
               blockchain,
               message.getData(),
-              ethereumWireProtocolConfiguration.getMaxGetBlockHeaders());
+              ethereumWireProtocolConfiguration.getMaxGetBlockHeaders().intValue());
       message.getPeer().send(response);
     } catch (final RLPException e) {
       LOG.debug(
@@ -110,7 +110,7 @@ class EthServer {
           constructGetBodiesResponse(
               blockchain,
               message.getData(),
-              ethereumWireProtocolConfiguration.getMaxGetBlockBodies());
+              ethereumWireProtocolConfiguration.getMaxGetBlockBodies().intValue());
       message.getPeer().send(response);
     } catch (final RLPException e) {
       LOG.debug(
@@ -126,7 +126,9 @@ class EthServer {
     try {
       final MessageData response =
           constructGetReceiptsResponse(
-              blockchain, message.getData(), ethereumWireProtocolConfiguration.getMaxGetReceipts());
+              blockchain,
+              message.getData(),
+              ethereumWireProtocolConfiguration.getMaxGetReceipts().intValue());
       message.getPeer().send(response);
     } catch (final RLPException e) {
       LOG.debug("Received malformed GET_RECEIPTS message, disconnecting: {}", message.getPeer(), e);
@@ -143,7 +145,7 @@ class EthServer {
           constructGetNodeDataResponse(
               worldStateArchive,
               message.getData(),
-              ethereumWireProtocolConfiguration.getMaxGetNodeData());
+              ethereumWireProtocolConfiguration.getMaxGetNodeData().intValue());
       message.getPeer().send(response);
     } catch (final RLPException e) {
       LOG.debug(
@@ -161,7 +163,7 @@ class EthServer {
           constructGetPooledTransactionsResponse(
               transactionPool,
               message.getData(),
-              ethereumWireProtocolConfiguration.getMaxGetPooledTransactions());
+              ethereumWireProtocolConfiguration.getMaxGetPooledTransactions().intValue());
       message.getPeer().send(response);
     } catch (final RLPException e) {
       LOG.debug(
